# frontend/generated/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "description": "Error response body, as produced by the AppError handler.\n\nReference it in endpoint ``responses={...}`` declarations so error\nstatuses are part of the OpenAPI contract (and the generated frontend\ntypes).",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": ["detail"],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatchCreate": {
        "additionalProperties": false,
        "description": "Payload for linking a session to a planned session by hand.",
        "properties": {
          "displaced": {
            "default": false,
            "title": "Displaced",
            "type": "boolean"
          },
          "planned_session_id": {
            "format": "uuid",
            "title": "Planned Session Id",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": ["session_id", "planned_session_id"],
        "title": "MatchCreate",
        "type": "object"
      },
      "MatchLinkStatus": {
        "description": "What one link between a recording and a planned session claims.\n\n``AUTO_HIGH`` and ``PENDING`` are machine verdicts and re-running matching\nmay replace either. ``CONFIRMED`` and ``DISPLACED`` are the athlete's and\nare **sticky**: :data:`STICKY_STATUSES` is the set matching never touches,\nwhich is what makes \"I already told you what this was\" hold across every\nlater re-run (build plan WP-6.6).\n\n``DISPLACED`` is the executed-instead-of link (WP-6.4): the athlete trained\nand it was not this, so the planned session is neither missed nor\ncompleted, and the activity is scored standalone with no adherence axes.",
        "enum": ["auto_high", "pending", "confirmed", "displaced"],
        "title": "MatchLinkStatus",
        "type": "string"
      },
...
      "Page_MatchRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MatchRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["items", "total", "offset", "limit"],
        "title": "Page[MatchRead]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "arc_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "arc",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/matches": {
      "get": {
        "description": "List match links, newest first. `?status=pending` is the proposal inbox.\n\nEvery row carries both sides, so the inbox renders and is answered without\na request per proposal.",
        "operationId": "matches-list_matches",
        "parameters": [
          {
            "description": "Restrict to one link status; omit for all of them. The proposal inbox is `pending`.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/MatchLinkStatus",
              "description": "Restrict to one link status; omit for all of them. The proposal inbox is `pending`."
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 2147483647,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MatchRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No valid session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Matches",
        "tags": ["matches"]
      },
      "post": {
        "description": "Link a session to a planned session by hand (build plan WP-6.6).\n\nAlways the athlete's own, and therefore **sticky**: no re-run of matching\nrevises or removes it. With `displaced` set it is the executed-instead-of\nlink \u2014 the planned session becomes `displaced` rather than completed, and\nthe activity is scored standalone.\n\nThe similarity is still computed and stored, however low: a deliberate link\nat 0.12 is worth being able to look at afterwards.",
        "operationId": "matches-create_match",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Malformed body"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No valid session"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No such session"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Already linked, or already ruled on"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorDetail"
                }
              }
            },
            "description": "The link violates a schema or domain rule"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Create Match",
        "tags": ["matches"]
      }
    },
    "/api/v1/matches/{match_id}": {
      "delete": {
        "description": "Unlink, putting both sides back exactly as they were (WP-6.8).\n\nExactly, not approximately: the link records the two statuses it displaced,\nso a session that was `displaced` before goes back to `displaced` and a\nplanned session that was `missed` goes back to `missed`.",
        "operationId": "matches-delete_match",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionMatchState"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No valid session"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No such match"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Delete Match",
        "tags": ["matches"]
      },
      "get": {
        "description": "One link, its score and the whole breakdown behind it.",
        "operationId": "matches-get_match",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No valid session"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No such match"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Match",
        "tags": ["matches"]
      },
      "patch": {
        "description": "Point a link at a different planned session (the swap, WP-6.6).\n\nOne operation rather than an unlink and a link, because it is one decision:\nthe old planned session goes back to exactly what it was, the new one takes\nthe link, and the result is confirmed \u2014 a retarget is always the athlete's.",
        "operationId": "matches-retarget_match",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRetarget"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Malformed body"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No valid session"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No such match"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Already linked, or already ruled on"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorDetail"
                }
              }
            },
            "description": "The link violates a schema or domain rule"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Retarget Match",
        "tags": ["matches"]
      }
    },
    "/api/v1/matches/{match_id}/confirm": {
      "post": {
        "description": "Accept a proposal (or an automatic link) as the athlete's own.\n\nThe session becomes `matched` and the planned session `completed`, and the\nlink stops being something matching may revise.",
        "operationId": "matches-confirm_match",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No valid session"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No such match"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Already linked, or already ruled on"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Confirm Match",
        "tags": ["matches"]
      }
    },
    "/api/v1/matches/{match_id}/reject": {
      "post": {
        "description": "Refuse a proposal: drop the link and call the session unplanned.\n\nNot the same as unlinking. Unlinking restores exactly what was there\nbefore; rejecting is the athlete saying \"that ride was not that session\",\nwhich leaves the ride `unplanned` and the planned session open for\nsomething else.",
        "operationId": "matches-reject_match",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionMatchState"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No valid session"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No such match"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Already linked, or already ruled on"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Reject Match",
        "tags": ["matches"]
      }
    },
...
    "/api/v1/sessions": {
      "get": {
        "description": "List completed sessions, newest first, optionally within a date range.\n\nA log rather than a calendar, so it reads backwards \u2014 the opposite of\n`GET /planned-sessions`, and deliberately: what happened is read from the\nmost recent, what is planned is read forwards.",
        "operationId": "sessions-list_sessions",
        "parameters": [
          {
            "description": "Earliest athlete-local date to include (inclusive).",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "description": "Earliest athlete-local date to include (inclusive).",
              "format": "date",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "description": "Latest athlete-local date to include (inclusive).",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "description": "Latest athlete-local date to include (inclusive).",
              "format": "date",
              "title": "End",
              "type": "string"
            }
          },
          {
            "description": "Restrict to one discipline; omit for all of them.",
            "in": "query",
            "name": "discipline",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SessionDiscipline",
              "description": "Restrict to one discipline; omit for all of them."
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 2147483647,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SessionListItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No valid session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Sessions",
        "tags": ["sessions"]
      }
    },
...
  }
}